requirements.txt
# LLM providers (install the one(s) you use)
groq>=0.9.0                  # Groq API (free tier)
google-generativeai>=0.8.0   # Google Gemini (free tier)
anthropic>=0.25.0             # Claude (paid, optional fallback)

# Core pipeline
streamlit>=1.35.0
pydantic>=2.0.0
jinja2>=3.1.0
weasyprint>=62.0
markdown>=3.6
python-dotenv>=1.0.0
pyyaml>=6.0
